fix: Cast calculator ID to integer to fix TypeError in trip sync
- Add (int) cast when reading FLESPI_TRIP_CALC_ID from config
- Fixes "Argument #2 ($calcId) must be of type int, string given" error
- Applied to both SyncFlespiTrips command and TripController

+++ app/Console/Commands/SyncFlespiTrips.php
@@ -40,7 +40,7 @@ class SyncFlespiTrips extends Command
      */
     public function handle(): int
     {
-        $calcId = config('services.flespi.trip_calculator_id');
+        $calcId = (int) config('services.flespi.trip_calculator_id');
 
         if (!$calcId) {
             $this->error('Trip calculator not configured. Please set FLESPI_TRIP_CALC_ID in .env file.');
